MODCXEKB-85 - Extract RMAPIService to a library module  * Splitted services
