+++ src/main/java/org/folio/holdingsiq/service/PackagesHoldingsIQService.java
@@ -1,0 +1,22 @@
+package org.folio.holdingsiq.service;
+
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.CompletionStage;
+
+import org.folio.holdingsiq.model.PackageByIdData;
+import org.folio.holdingsiq.model.PackageId;
+import org.folio.holdingsiq.model.PackagePost;
+import org.folio.holdingsiq.model.PackagePut;
+import org.folio.holdingsiq.model.Packages;
+import org.folio.holdingsiq.model.Sort;
+
+public interface PackagesHoldingsIQService {
+
+  CompletableFuture<PackageByIdData> retrievePackage(PackageId packageId);
+  CompletableFuture<Packages> retrievePackages(Long providerId);
+  CompletableFuture<Packages> retrievePackages(
+    String filterSelected, String filterType, Long providerId, String q, int page, int count, Sort sort);
+  CompletableFuture<PackageByIdData> postPackage(PackagePost entity, Long vendorId);
+  CompletionStage<Void> updatePackage(PackageId packageId, PackagePut packagePut);
+  CompletableFuture<Void> deletePackage(PackageId packageId);
+}
