+++ src/main/java/org/folio/holdingsiq/service/ResourcesHoldingsIQService.java
@@ -1,0 +1,18 @@
+package org.folio.holdingsiq.service;
+
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.CompletionStage;
+
+import org.folio.holdingsiq.model.ResourceId;
+import org.folio.holdingsiq.model.ResourcePut;
+import org.folio.holdingsiq.model.ResourceSelectedPayload;
+import org.folio.holdingsiq.model.Title;
+
+public interface ResourcesHoldingsIQService {
+
+  CompletableFuture<Title> postResource(ResourceSelectedPayload resourcePost, ResourceId resourceId);
+  CompletionStage<Void> updateResource(ResourceId parsedResourceId, ResourcePut resourcePutBody);
+  CompletableFuture<Void> deleteResource(ResourceId parsedResourceId);
+  CompletableFuture<Title> retrieveResource(ResourceId resourceId);
+
+}
